Clean up int -> long ID conversion
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@20356 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Controller/src/main/java/org/gusdb/wdk/controller/action/GetFilterSummaryAction.java
@@ -38,7 +38,7 @@ public class GetFilterSummaryAction extends Action {
       throw new WdkUserException("Required step parameter is missing.");
     
     UserBean user = ActionUtility.getUser(servlet, request);
-    StepBean step = user.getStep(Integer.valueOf(stepId));
+    StepBean step = user.getStep(Long.valueOf(stepId));
     AnswerValueBean answer = step.getAnswerValue();
     Filter filter = answer.getQuestion().getFilter(filterName);
     FilterSummary summary = answer.getFilterSummary(filterName);
